Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/selectors/matches-nested-expected.txt

Yellow
Red

FAIL Test nested :matches() chooses highest specificity for class outside :matches(). assert_equals: expected "20px" but got "10px"
FAIL Test nested :matches() specificity for class within arguments. assert_equals: expected "rgb(255, 0, 0)" but got "rgb(0, 0, 0)"
